+++ instrumentation/apache-httpclient-4.0/src/main/java/io/opentelemetry/instrumentation/hypertrace/apachehttpclient/v4_0/ApacheHttpClientObjectRegistry.java
@@ -14,16 +14,15 @@
  * limitations under the License.
  */
 
-package io.opentelemetry.instrumentation.hypertrace.apache.httpclient;
+package io.opentelemetry.instrumentation.hypertrace.apachehttpclient.v4_0;
 
-public class ApacheHttpClientInstrumentationName {
-  public static final String PRIMARY = "httpclient";
-  public static final String[] OTHER = {
-    "apache-httpclient",
-    "apache-http-client",
-    "ht",
-    "httpclient-ht",
-    "apache-httpclient-ht",
-    "apache-http-client-ht"
-  };
+import com.blogspot.mydailyjava.weaklockfree.WeakConcurrentMap;
+import io.opentelemetry.api.trace.Span;
+
+public class ApacheHttpClientObjectRegistry {
+
+  private ApacheHttpClientObjectRegistry() {}
+
+  public static final WeakConcurrentMap<Object, Span> objectToSpanMap =
+      new WeakConcurrentMap<>(false);
 }
